src/5. Tools and Agents/3_multi_agent_systems/3_Google_A2A_protocol/standalone agents/A2A_standalone_agent_wikipedia/agent_tests.py
from agent import WikipediaAgent
from dotenv import load_dotenv
load_dotenv()

agent = WikipediaAgent()
sessionID = "session_12345"

query = "Tell me abaout Adam Mickiewicz"
response = agent.invoke(query, sessionID)
print(response)